requirements.txt
Flask
Pillow
werkzeug